+++ etna/transforms/decomposition/__init__.py
@@ -17,4 +17,5 @@ from etna.transforms.decomposition.deseasonal import DeseasonalityTransform
 from etna.transforms.decomposition.detrend import LinearTrendTransform
 from etna.transforms.decomposition.detrend import TheilSenTrendTransform
 from etna.transforms.decomposition.dft_based import FourierDecomposeTransform
+from etna.transforms.decomposition.model_based import ModelDecomposeTransform
 from etna.transforms.decomposition.stl import STLTransform