+++ tests/test_transforms/test_inference/test_inverse_transform.py
@@ -7,6 +7,7 @@ from ruptures import Binseg
 from sklearn.tree import DecisionTreeRegressor
 
 from etna.analysis import StatisticsRelevanceTable
+from etna.models import HoltWintersModel
 from etna.models import ProphetModel
 from etna.transforms import AddConstTransform
 from etna.transforms import BinaryOperationTransform
@@ -47,6 +48,7 @@ from etna.transforms import MedianTransform
 from etna.transforms import MinMaxDifferenceTransform
 from etna.transforms import MinMaxScalerTransform
 from etna.transforms import MinTransform
+from etna.transforms import ModelDecomposeTransform
 from etna.transforms import MRMRFeatureSelectionTransform
 from etna.transforms import OneHotEncoderTransform
 from etna.transforms import PredictionIntervalOutliersTransform
@@ -144,6 +146,7 @@ class TestInverseTransformTrain:
                 {},
             ),
             (FourierDecomposeTransform(in_column="target", k=5, residuals=True), "regular_ts", {}),
+            (ModelDecomposeTransform(model=ProphetModel(), in_column="target", residuals=True), "regular_ts", {}),
             # embeddings
             (
                 EmbeddingSegmentTransform(
@@ -607,6 +610,7 @@ class TestInverseTransformTrain:
                 {},
             ),
             (FourierDecomposeTransform(in_column="target", k=5, residuals=True), "regular_ts", {}),
+            (ModelDecomposeTransform(model=HoltWintersModel(), in_column="target", residuals=True), "regular_ts", {}),
             # embeddings
             (
                 EmbeddingSegmentTransform(
@@ -1100,6 +1104,7 @@ class TestInverseTransformTrainSubsetSegments:
                 "regular_ts",
             ),
             (FourierDecomposeTransform(in_column="target", k=5, residuals=True), "regular_ts"),
+            (ModelDecomposeTransform(model=HoltWintersModel(), in_column="target", residuals=True), "regular_ts"),
             # embeddings
             (
                 EmbeddingSegmentTransform(
@@ -1388,6 +1393,8 @@ class TestInverseTransformFutureSubsetSegments:
             ),
             (FourierDecomposeTransform(in_column="target", k=5, residuals=True), "regular_ts"),
             (FourierDecomposeTransform(in_column="positive", k=5, residuals=True), "ts_with_exog"),
+            (ModelDecomposeTransform(model=HoltWintersModel(), in_column="target", residuals=True), "regular_ts"),
+            (ModelDecomposeTransform(model=HoltWintersModel(), in_column="positive", residuals=True), "ts_with_exog"),
             # embeddings
             (
                 EmbeddingSegmentTransform(
@@ -1980,6 +1987,7 @@ class TestInverseTransformTrainNewSegments:
                 ),
                 "regular_ts",
             ),
+            (ModelDecomposeTransform(model=ProphetModel(), in_column="target", residuals=True), "regular_ts"),
             # encoders
             (MeanEncoderTransform(in_column="weekday", out_column="mean_encoder"), "ts_with_exog"),
             (MeanSegmentEncoderTransform(), "regular_ts"),
@@ -2421,6 +2429,7 @@ class TestInverseTransformFutureNewSegments:
                 ),
                 "regular_ts",
             ),
+            (ModelDecomposeTransform(model=ProphetModel(), in_column="target", residuals=True), "regular_ts"),
             # encoders
             (MeanEncoderTransform(in_column="weekday", out_column="mean_encoder"), "ts_with_exog"),
             (MeanSegmentEncoderTransform(), "regular_ts"),
@@ -3006,6 +3015,7 @@ class TestInverseTransformFutureWithTarget:
         "transform, dataset_name, expected_changes",
         [
             (FourierDecomposeTransform(in_column="target", k=5, residuals=True), "regular_ts", {}),
+            (ModelDecomposeTransform(model=HoltWintersModel(), in_column="target", residuals=True), "regular_ts", {}),
         ],
     )
     def test_inverse_transform_future_with_target_fail_require_history(
@@ -3127,6 +3137,8 @@ class TestInverseTransformFutureWithoutTarget:
             ),
             (FourierDecomposeTransform(in_column="target", k=5, residuals=True), "regular_ts", {}),
             (FourierDecomposeTransform(in_column="positive", k=5, residuals=True), "ts_with_exog", {}),
+            (ModelDecomposeTransform(model=ProphetModel(), in_column="target", residuals=True), "regular_ts", {}),
+            (ModelDecomposeTransform(model=ProphetModel(), in_column="positive", residuals=True), "ts_with_exog", {}),
             # embeddings
             (
                 EmbeddingSegmentTransform(